requirements.txt
Flask==2.0.1
line-bot-sdk==1.19.0
requests==2.26.0
gunicorn==20.1.0
werkzeug==2.0.3
google-generativeai==0.6.0
google-cloud-dialogflow==2.14.1
google-api-python-client==2.30.0
google-auth==2.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
